Just Humberto fix
It seems update_target only actually updates targets if it wants to and the previous code was written expecting this function to be called on every global step. Now it allows being called later than expected. Of course this comes at a price of not actually performing the update at quite the frequency given in config.. (/sad-trombone)